Fix merge problems. Fixes #19
