+++ src/core/parser.py
@@ -1,16 +1,21 @@
 from tools.cmp.evaluation import evaluate_reverse_parse
-from tools.cmp.pycompiler import Grammar
 from tools.cmp.utils import Token
 from tools.parsers import LR1Parser
 from core.grammar import G
 
-class Parser:
-  def __init__(self, G: Grammar) -> None:
-    self.parse = LR1Parser(G=G, verbose=False)
+parser = None
 
-  def parse(tokens: list[Token]):
-    pass
-    # parse(w=List[Token], shift-reduce=True)
-    # recibir el ast y aplicar evaluate_reverse_parse
-    # return ast
+def parse(tokens: list[Token]):
+    print('>>> Parsing...')
+
+    if parser == None:
+        parse = LR1Parser(G,verbose=False)
+    result = parse(tokens, get_shift_reduce=True)
+
+    right_parse, operations = result
+    ast = evaluate_reverse_parse(right_parse, operations, tokens)
+    return ast
+
+def build_parser():
+    return LR1Parser(G, verbose=False)
