--- src/main.py
@@ -1,5 +1,0 @@
-from tools.cmp.pycompiler import Grammar
-
-from tools.ast import *
-from tools.Lexer.lexer import Lexer
-from tools.parsers import LR1Parser
