+++ tests/test.1.hulk
@@ -1,0 +1,35 @@
+ function tan(x) => sin(x) / cos(x);
+ function cot(x) => 1 / tan(x);
+ function tan(x) => sin(x) / cos(x);
+  
+function operate(x, y) {
+    print(x + y);
+    print(x - y);
+    print(x * y);
+    print(x / y);
+}
+function operate(x, y) {
+    print(x + y);
+    print(x - y);
+    print(x * y);
+    print(x / y);
+ };
+
+type Point {
+    x = 0;
+    y = 0;
+
+    getX() => self.x;
+    getY() => self.y;
+    setX(x) => self.x[0] := x;
+    setY(y) => self.y[int(sin(self.y))] := y;
+ }
+
+ let array =[1,2,3,4] in
+    let a = [x^(sen(4)+3) || x in array ] in {
+        if(a.length ==0){
+            a:= new Point(0,4);
+            a.getX();
+        }
+        else a.getX();
+    }
